kayam-portfolio/requirements.txt
pdfplumber>=0.7.0
pytesseract>=0.3.10
Pillow>=9.0.0
pandas>=1.5.0